Removed references to base project
Any place the term 'project' is used, it implies a reference to a Fabric canvas wrapper that holds some getters for the Fabric canvas properties. I think I will probably add a setContext function so this object can reference a canvas directly. Eventually, maybe we should convert this to a class object so a Fabric context object could be passed in a constructor.